Added classes: UserDetailsPage, UserDetailsPageMap, UserDetailsPageValidator
