+++ TeamProjectFluxday/TeamProjectFluxday/Pages/Users/UserDetails/UserDetailsPageValidator.cs
@@ -1,0 +1,8 @@
+﻿using TeamProjectFluxday.Core;
+
+namespace TeamProjectFluxday.Pages
+{
+    public class UserDetailsPageValidator : BasePageValidator<UserDetailsPageMap>
+    {
+    }
+}
