+++ TeamProjectFluxday/TeamProjectFluxday/Tests/UsersPageTests.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Threading;
 using TeamProjectFluxday.Core;
 using TeamProjectFluxday.Data;
 using TeamProjectFluxday.Pages;
@@ -24,7 +25,7 @@ namespace TeamProjectFluxday.Tests
         public void Test001UsersPageValidation()
         {
             var usersPage = dashboardPage.NavigationPanel.NavigateToUsersPage();
-
+            Thread.Sleep(2000);
             usersPage.Validate().IsOnUsersPage();
         }
     }
